requirements.txt
aioredis==1.2.0
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
autopep8==1.4.4
bcrypt==3.1.6
blessings==1.7
bpython==0.18
capstone==4.0.1
certifi==2019.3.9
cffi==1.12.2
channels==1.1.7
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.6.1
curtsies==0.3.0
daphne==1.3.0
Django==2.1.7
django-appconf==1.0.3
django-avatar==4.1.0
django-console==0.4.7
django-debug-toolbar==1.11
django-filter==2.1.0
django-model-utils==3.1.2
django-multiselectfield==0.1.8
django-simple-gravatar==1.0.1
djangorestframework==3.9.4
docopt==0.6.2
filelock==3.0.10
future==0.17.1
greenlet==0.4.15
h2==3.1.0
hiredis==1.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
intervaltree==2.1.0
isort==4.3.18
jedi==0.13.3
lazy-object-proxy==1.4.0
Mako==1.0.8
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
msgpack-python==0.5.6
packaging==19.0
paramiko==2.4.2
parso==0.4.0
Pillow==6.0.0
pluggy==0.9.0
priority==1.3.0
prompt-toolkit==1.0.14
psutil==5.6.1
pwn==1.0
pwntools==3.12.2
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==3.0.0
pyelftools==0.25
pyfiglet==0.8.post1
pyflakes==2.1.1
Pygments==2.3.1
PyHamcrest==1.9.0
PyInquirer==1.0.3
pylint==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pypandoc==1.4
pyparsing==2.3.1
pyserial==3.4
PySocks==1.6.8
python-dateutil==2.8.0
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
pytz==2018.9
redis==2.10.6
regex==2019.4.14
requests==2.21.0
rope==0.14.0
ROPGadget==5.4
service-identity==18.1.0
six==1.12.0
snowballstemmer==1.2.1
sortedcontainers==1.5.10
sqlparse==0.3.0
toml==0.10.0
tox==3.8.6
Twisted==19.2.0
txaio==18.8.1
typed-ast==1.3.5
typing==3.6.6
unicorn==1.0.1
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
wrapt==1.11.1
yapf==0.27.0
zope.interface==4.6.0
